requirements.txt
Flask
praw
PyYAML
wsgi
pony
markdown
bs4
python-Levenshtein
wheel

#nginx unit and unit-python are also needed to run the webapp